// src/app/shared/Models/GetClient.ts
export class GetClient {
    clientId:string;
    logoPath: string;
    cityId: number;
    governorateId: number;
    cityTitle:string;
    governorateTitle:string;
    clientTypeTitle: string;
    clientTypeId: number;
    name: string;
    mobile: string;
    address: string;
    userName: string;
    password:string
}